refactor[#23] : 이메일 회원가입 폼 파일 이름 수정
