+++ src/features/auth/ui/index.ts
@@ -1,9 +1,8 @@
 export * from "./hyperlinks";
-export { default as SignUpByEmailForm } from "./SignUpByEmailForm";
+export * from "./signUpByEmailForm";
 export * as LoginForm from "./LoginForm";
 export * from "./petInformationForm";
 export { default as UserInfoRegistrationForm } from "./UserInfoRegistrationForm";
 export * from "./authNavigationBar";
 export * from "./regionModal";
 export * from "./dashboard";
-
