+++ src/features/auth/ui/signUpByEmailForm.tsx
